Calculation and plot functions in the wibs package

This function does nothing but collect the shared @para in multiple calculation and plot functions.

mean_calc_df() calculates mean values for a data frame by groups

mean_calc_files() calculates time series of mean values from multiple *_Particle_preprocessed.csv files

dp_break() creates particle diameter groups

dist_calc() calculates the size distribution.

dist_calc_type() calculates the size distribution grouped by type

dist_calc_type_files() calculates the size distribution of multiple ParticleData_preprocessed.csv files

dist_calc_ts() calculates the time series of the distribution

dist_calc_ts_files() calculates the size distribution of multiple ParticleData_preprocessed.csv files

read_csvs() reads multiple csv files with the same column names.

summary_plot_all() plots all the summary data from multiple _summary.csv files in one image.

mean_plot_df() plots the time series of mean values from a dataframe.

mean_plot_files() plots the time series of mean values in multiple "_mean.csv" files

mean_plot_stacked_df() plots a stacked area plot of time series

mean_plot_stacked_files() plots area plots of time series from multiple *_mean.csv files

mean_plot_stacked_all() plots area plots of time series from *_mean.csv files

dist_plot_type() plots the size distribution based on the data calculated by dist_calc_type()

dist_plot_type_files() plots the distribution images from "_Ndist_all.csv" files

dist_plot_ts_df() plots the time series of the distribution

dist_plot_ts_files() plots the distribution time series images from "Ndist_ts_.csv" files

dist_plot_ts_all() plots all the 3D time series of the size distribution from multiple csv. files in one image.

abc_plot_venn_df() plots a Venn diagram for A B C.

abc_plot_venn_df() plots a Venn diagram for A B C.

abc_plot_venn_all() plots area plots of time series from *_mean.csv files

mean_calc_df(dtf, column_names, myfactors, method = "mean", savefile_folder = NA,
    savefile_suffix = "", v_factor = 1)

mean_calc_files(fns, minute_bin, v_factor, myfactors)

dp_break(size_range, size_bin)

dist_calc(size, size_range, size_bin, v_factor = 1, savefile_folder = NA,
    savefile_suffix = "")

dist_calc_type(size, type, size_range, size_bin, v_factor, savefile_folder = NA,
    savefile_suffix = "")

dist_calc_type_files(fns, size_range, size_bin)

dist_calc_ts(size, time, size_range, size_bin, v_factor, savefile_folder = NA,
    savefile_suffix = "")

dist_calc_ts_files(fns, size_range, size_bin, v_factor, minute_bin)

read_csvs(fns, add_id = FALSE)

summary_plot_all(fns, savefile_folder = ".", savefile_suffix = "", image_format = "png")

mean_plot_df(dtf, column_names, savefile_folder = NA, savefile_suffix = "",
    image_format = "png")

mean_plot_files(fns, column_names = c("count", "Size_um", "Xe1_FluorPeak_1",
    "Xe1_FluorPeak_2", "Xe2_FluorPeak_2", "Asphericity"), savefile_suffix = "",
    image_format = "png")

mean_plot_stacked_df(dtf, prefix, ylab, savefile_folder = NA, savefile_suffix = "",
    image_format = "png", keyword = c("A", "B", "C", "AB", "AC", "BC", "ABC"))

mean_plot_stacked_files(fns, prefix, ylab, savefile_suffix = "", image_format = "png",
    keyword = c("A", "B", "C", "AB", "AC", "BC", "ABC"))

mean_plot_stacked_all(fns, prefix = "count", ylab, savefile_folder = NA,
    savefile_suffix = "", image_format = "png", keyword = c("A", "B", "C",
        "AB", "AC", "BC", "ABC"))

dist_plot_type(dtf, savefile_folder = NA, savefile_suffix = "", image_format = "png")

dist_plot_type_files(fns, image_format = "png")

dist_plot_ts_df(dtf, savefile_folder = NA, savefile_suffix = "", image_format = "png",
    if_legend = TRUE)

dist_plot_ts_files(fns, savefile_suffix = "", image_format = "png")

dist_plot_ts_all(fns, ylab, savefile_folder = ".", savefile_suffix = "",
    image_format = "png")

abc_plot_venn_df(dtf, savefile_folder = ".", savefile_suffix = "", image_format = "png")

abc_plot_venn_files(fns, image_format = "png")

abc_plot_venn_all(fns, savefile_folder = NA, savefile_suffix = "", image_format = "png")

Arguments

dtf
data frame as input. can be:

dist_data, calculated by dist_calc_type() or dist_calc_ts().

pdata particle data frame with 'time_binned' and 'type' columns

pmean the dataframe of particle mean values. with a column 'time'.

column_names
the names of the columns to calcualte or plot.

myfactors
the names of the columns as groups for tapply calculation. Here it is c('time_binned', 'type').

method
the function name, such as 'sum' or 'mean'

savefile_folder
the folder name for writing the calculated results. if NA, no writing.

savefile_suffix
the suffix of the data file name, such as minute_bin

v_factor
air volume within a time bin. if volume involved (method 'sum'), v_factor = minute_bin * flowrate. If the bulk calculation for an entire measurement, v_factor is the total volumn. othewise (method 'mean'), v_factor = 1.

fns
character.

For mean_calc_files(), dist_calc_type_files(), dist_calc_ts_files(): The paths of the *_ParticleData_preprocessed.csv files

For mean_plot_files(), mean_plot_stacked_files(), mean_plot_stacked_all() : The paths of the *_mean.csv files

For dist_plot_type_files(): The paths of the *_Ndist_all.csv files

For dist_plot_ts_files(), dist_plot_ts_all(): "Ndist_ts_.csv" files

minute_bin
time bin in minute

size_range
the range of Dp

size_bin
how many bins

size
vector. the diameter data

type
vector. aerosol type data

time
vector. the grouped time (time_binned).

add_id
logical. whether add a column of id from the directory name

image_format
without '.'

prefix
the prefix of the selected column name

ylab
ylab

keyword
keyword for picking out column names

if_legend
logical. Whether to display legend.

abc_n
datafame. The number of A, B, C

Value

mean_calc_df() returns a dataframe with the mean values

mean_calc_files() returns files of time series of mean

dp_break() returns a list.

dist_calc() returns a data frame or a file of dN/dlogDp vs Dp.

dist_calc_type() returns a dataframe or a file.

dist_calc_type_files() returns files of dN/dlogDp vs Dp.

dist_calc_ts() returns a datafram of a file.

dist_calc_ts_files() returns files of time series of dN/dlogDp

read_csvs(): a data frame

summary_plot_all(): a summary csv file and a figure file.

mean_plot_df() returns image files.

mean_plot_files() returns image files

mean_plot_stacked_df() returns a figure file.

mean_plot_stacked_files() returns image files.

mean_plot_stacked_all(): image files.

dist_plot_type(): a figure (file).

dist_plot_type_files(): image files.

dist_plot_ts_df(): a figure (file)

dist_plot_ts_files(): image files

dist_plot_ts_all(): a figure file.

abc_plot_venn_df(): a figure file.

abc_plot_venn_df(): a figure file.

abc_plot_venn_all(): image files.